Menu, Materia, Cambios de Contra
Carga, modificacion y eliminacion de materias, Creacion de primera contraseña, Cambio de contraseña. Tuve que agregar unas cosas al menu principal y al ingreso.

+++ Sistema/DAL/DalMateria.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using System.Collections.Generic;
 using System.Text;
 using Sistema.DAL;
-
+using System.Data;
 namespace Sistema.DAL
 {
     class DalMateria
@@ -25,17 +25,61 @@ namespace Sistema.DAL
                 return false;
             }
         }
+        public bool ModificarMateria (int idMateria,int idCarrera, string nombre, int IDProfesor, int año, int horas, int Promocional)
+        {
+            string Consulta = "update Materia set CarreraID="+idCarrera+",ProfesorID="+IDProfesor+",Nombre='"+nombre+"',Año="+año+",CargaHoraria="+horas+",Promocional="+Promocional+" where MateriaID="+idMateria;
+            try
+            {
+                ConexionBD.Actualizar(Consulta);
+                return true;
+            }
+            catch (Exception)
+            {
+                return false;
+            }
+        }
          public bool EliminarMateria (int IDMateria)
         {
             try
             {
-                ConexionBD.Eliminar("delete from Materia where MateriaID =" + IDMateria );
+                DataSet DS = ConexionBD.consultar("select * from Correlatividades where MateriaID=" + IDMateria + " or MateriaIDp =" + IDMateria);
+                if(DS.Tables[0].Rows.Count == 0) {
+                    try
+                    {
+                        ConexionBD.Eliminar("delete from Materia where MateriaID =" + IDMateria);
+                        return true;
+
+                    }
+                    catch (Exception)
+                    {
+
+                        return false;
+                    }
+ 
+                }
+                ConexionBD.Eliminar("delete from Correlatividades where MateriaID =" + IDMateria + " or MateriaIDp =" + IDMateria);
+                ConexionBD.Eliminar("delete from Materia where MateriaID =" + IDMateria);
                 return true;
             }
-            catch (Exception ex)
+            catch (Exception)
             {
                 return false;
+                 
+
             }
+         //   try
+          //  {
+                //ConexionBD.Begin();
+          //      ConexionBD.Eliminar("delete from Correlatividades where MateriaID =" + IDMateria +" or MateriaIDp ="+ IDMateria);
+           //     ConexionBD.Eliminar("delete from Materia where MateriaID =" + IDMateria );
+               // ConexionBD.Commit();
+         
+           //     return true;
+           // }
+           // catch (Exception ex)
+           // {       // ConexionBD.Rollback();
+            //    return false;
+     //       }
         }
     }
 }
